requirements.txt
anthropic
modal
numpy
openai
packaging
pydra_config
torch
tqdm
datasets
transformers
google-generativeai
together
ninja